Flesh out profile features and add profile error codes

+++ ProjectSparrow/Web/Models/UtilityModels.cs
@@ -7,8 +7,8 @@ namespace Web.Models.Utilities
 {
     public class GeoLocation
     {
-        public double Latitude { get; set; }
+        public float Latitude { get; set; }
 
-        public double Longitude { get; set; }
+        public float Longitude { get; set; }
     }
 }
